% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.R
\name{has_names}
\alias{has_names}
\title{Checks if a list has names.}
\usage{
has_names(dat)
}
\arguments{
\item{dat}{list. The list to verify.}
}
\value{
boolean. TRUE if the list is named, FALSE otherwise.
}
\description{
Checks if a list has names.
}

