% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.R
\name{is_heterogeneous}
\alias{is_heterogeneous}
\title{Is this heterogeneous data?}
\usage{
is_heterogeneous(dat)
}
\arguments{
\item{dat}{list. The list to verify.}
}
\value{
boolean. TRUE if the list is heterogeneous, FALSE otherwise.
}
\description{
Is this heterogeneous data?
}

