% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.R
\name{is_homogeneous}
\alias{is_homogeneous}
\title{Is this homogeneous data?}
\usage{
is_homogeneous(dat)
}
\arguments{
\item{dat}{list. The list to verify.}
}
\value{
boolean. TRUE if the list is heterogeneous, FALSE otherwise.
}
\description{
Is this homogeneous data?
}

