% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.recombinator.R
\docType{package}
\name{recombinator}
\alias{recombinator}
\alias{recombinator-package}
\title{Turn nested lists into data.frames.}
\description{
A mini-utility package for turning nested lists into data.frames.
}

