% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombinator.R
\name{warn_on_nonstandard_names}
\alias{warn_on_nonstandard_names}
\title{Warn if names will be changed when converting to a data.frame.}
\usage{
warn_on_nonstandard_names(data)
}
\arguments{
\item{data}{list. A list to convert to a data.frame.}
}
\value{
Nothing, but a \code{warning} if the names will be
   mangled due to R's \code{\link{make.names}}.
}
\description{
Warn if names will be changed when converting to a data.frame.
}

